requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.1
amqp==5.1.1
anyio==3.6.2
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
autoflake==2.0.2
backcall==0.2.0
bcrypt==4.0.1
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
comm==0.1.2
debugpy==1.6.6
decorator==5.1.1
dnspython==2.3.0
ecdsa==0.18.0
email-validator==1.3.1
exceptiongroup==1.1.1
executing==1.2.0
fastapi==0.92.0
fastapi-cache2==0.2.1
fastapi-versioning==0.10.0
flake8==6.0.0
flower==1.2.0
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanize==4.6.0
idna==3.4
importlib-metadata==6.0.0
iniconfig==2.0.0
ipykernel==6.21.3
ipython==8.11.0
isort==5.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.0.3
jupyter_core==5.2.0
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nodeenv==1.7.0
orjson==3.8.6
packaging==23.0
parso==0.8.3
passlib==1.7.4
pathspec==0.11.1
pendulum==2.1.2
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.1
pluggy==1.0.0
prometheus-client==0.16.0
prometheus-fastapi-instrumentator==5.11.2
prompt-toolkit==3.0.38
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.5
pyflakes==3.0.1
Pygments==2.14.0
pyright==1.1.300
pytest==7.2.2
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-json-logger==2.0.7
python-multipart==0.0.5
pytz==2022.7.1
pytzdata==2020.1
pywin32==305; platform_system=="Windows"
PyYAML==6.0
pyzmq==25.0.0
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
sentry-sdk==1.17.0
six==1.16.0
sniffio==1.3.0
sqladmin==0.10.0
SQLAlchemy==2.0.5.post1
stack-data==0.6.2
starlette==0.25.0
tomli==2.0.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.20.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
WTForms==3.0.1
yarl==1.9.2
zipp==3.15.0
